auto-detect column datatypes for the object store to provide proper filtering, search and sorting based on the column type.
Raw Data column type leaves the data as it was received from the store (so it can be of mixed types for example). Sorting, filtering and search happens after casting the raw data to a string though.

For other datatypes, if a value does not match the column type, then it is set as undefined. null is left as is though (given it usually means a missing value rather than an invalid one)