move the logic for auto-detecting the datatypes to the inspected window
That's because only json-serializable data can be passed between the inspected window and the extension. So, we detect the types and serialize dates/bigints as strings to be able to later convert them back. Unsupported datatypes (maps, sets, ...) are set as null to avoid chrome.devtools.inspectedWindow.eval erroring out when trying to pass the data
https://developer.chrome.com/docs/extensions/reference/api/devtools/inspectedWindow#:~:text=The%20evaluated%20code%20may%20return%20a%20value%20that%20is%20passed%20to%20the%20extension%20callback.%20The%20returned%20value%20has%20to%20be%20a%20valid%20JSON%20object%20(it%20may%20contain%20only%20primitive%20JavaScript%20types%20and%20acyclic%20references%20to%20other%20JSON%20objects).